Finish Knight Travails project
Created knightMoves function which uses an adjacency matrix to mark
visited spots so we don't revisit and a queue to keep track of which
nodes need to be searched that are next in the level(if matching the
target move we don't have to search any shorter paths because it's the
shortest).

Added "index.html" file to access the debugger.